agents/shared/requirements.txt
langgraph>=0.0.10
google-genai>=1.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
cuid2>=2.0.0
structlog>=24.0.0
redis>=5.0.0
opentelemetry-api>=1.24.0
